### 2021-02-18 #103 @whedon
[PRE REVIEW]: WI Fast Stats: a collection of web apps for the visualization and analysis of WI Fast Plants data
**Submitting author:** @crsl4 (<a href="http://orcid.org/0000-0002-9789-8915">Claudia Solis-Lemus</a>)
**Repository:** <a href="https://github.com/crsl4/fast-stats" target ="_blank">https://github.com/crsl4/fast-stats</a>
**Version:** v2.0
**Reviewers:** pierrerogy, magitz

@crsl4 if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
