### 2021-02-18 #103 @labarba
hi there! 👋

We've been slow here, due to unmanageable work loads that trump volunteer tasks. Thanks for your patience. I note that your paper did not compile, but inspected the [markdown file for the article](https://github.com/crsl4/fast-stats/blob/master/ms/paper.md). It is not clear how this work fits in the scope of JOSE, as this seems to be a pure web app?

JOSE aims to publish computational learning modules, where computing (coding) is used in a context of learning another subject. We extend the scope to allow for software infrastructure in support of computational learning, e.g., auto-graders of student-submitted code. But we did not target JOSE for web apps that learners interact with as a GUI. If you wanted your submission considered regardless of this, you would make your case in the paper. See the section on our documentation about [what the paper should contain](https://openjournals.readthedocs.io/en/jose/submitting.html#what-should-my-paper-contain).
