### 2021-02-18 #103 @crsl4
Thanks @labarba ! I completely understand the reviewing speed.

We originally submitted the paper to JOSS because this is an open-source bioinformatics tool (collection of web apps) that students and researchers can use for data visualization and data analysis on plant experimental data that is very widely used across the US ([WI Fast Plants](https://fastplants.org/)). For example, experiments with WI Fast Plants have been run by K-16 teachers around the world for nearly 30 years as an educational model-organism, so we have 30 years of collected data with no tools to visualize/analyze! We produced this suite as a data science tool to researchers/students doing research on WI Fast Plants so that 1) they have tools to visualize/analyze, 2) they can explore how the web apps are created by browsing the code. Existing data science tools are proprietary, so they are just a black box!

Anyway, personally I think that it fits well with JOSS given that it is an open-source bioinformatics tool, but given that our audience are K-16 teachers and students doing research, the editors at JOSS suggested that we submit to JOSE.

So, I am not sure what is the best course of action!
Back to the JOSE scope, we are providing a software infrastructure in support of learning, but the learning is on the plants based on collected data from experiments, we are just providing the bioinformatics tool.

Let me know what you think, thanks!
