+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(fast-stats)
+
+test_check("fast-stats")
