### 2021-02-21 #103 @labarba
I see. Thanks for the clarification. Internally, JOSE editors have been doing some soul-searching on our scope for web applications, but we have published tools of that category. I understand that JOSS would not consider it, not being research software per se.

Question: are data sets also open? That is, would a reviewer be able to test the apps with real data that you provide?
