### 2021-02-22 #103 @crsl4
Yes! There are open data in the web apps.
The goal for the web apps is that people will collect their own data from the plant experiments and then use the web app to analyze it.
But the web apps have built in sample data that anyone can use to learn about the plant experiments without having to do the plant experiments themselves.
Using the test data is as simple as clicking the "sample data" option in the Data Upload section.
Currently, the web apps will be used in the Intro Biology class at our university (with 1400 students) who will conduct research experiments with fast plants and use the app to analyze the collected data. We think it is a very useful software, but at the same time, we completely understand if it does not fall within the scope of JOSE (even after the soul-searching ;) )
