+++ shiny-app/webinar-bio152/server.R
@@ -1,0 +1,1244 @@
+library(data.table)
+library(ggplot2)
+library(car)
+library(plotly)
+library(pracma)
+library(graphics)
+library(ggplotify)
+library(thatssorandom)
+library(rsconnect)
+library(viridis)
+library(RColorBrewer)
+# library(shinyjs)
+# Define server logic to read selected file ----
+# global variable
+options(shiny.maxRequestSize=10*1024^2)
+toy<-read.csv("toys.csv",
+              header = T,
+              sep = ",")
+server <- function(input, output,session) {
+  # #################################### 
+  dsnames<-c() #a vector to store col names
+  
+  data_set <- reactive({
+    if(is.null(input$file1)) return(NULL)
+   
+    req(input$file1)
+    tryCatch(
+      {
+        df <- read.csv(input$file1$datapath,
+                       header = input$header,
+                       sep = input$sep)
+      },
+      error = function(e) {
+        # return a safeError if a parsing error occurs
+        stop(safeError(e))
+      }
+    )
+  })
+  output$fileUploaded <- reactive({
+    return(!is.null(data_set()))
+  })
+  outputOptions(output, 'fileUploaded', suspendWhenHidden=FALSE)
+  observe({
+    if(input$fileType=="uploadFile"){
+      dsnames <- names(data_set())
+      cb_options <- list()
+      cb_options[ dsnames] <- dsnames
+      updateRadioButtons(session, "xaxisGrp",
+                         label = "Group Variable",
+                         choices = cb_options,
+                         selected = "")
+      updateRadioButtons(session, "yaxisGrp",
+                         label = "Quantity",
+                         choices = cb_options,
+                         selected = "")
+      updateRadioButtons(session, "groupVar",
+                         label = "Group Variable",
+                         choices = cb_options,
+                         selected = "")
+      updateRadioButtons(session, "quantity",
+                         label = "Quantity",
+                         choices = cb_options,
+                         selected = "")
+      updateRadioButtons(session, "gv1",
+                         label = "Group Variable 1",
+                         choices = cb_options,
+                         selected = "")
+      updateRadioButtons(session, "gv2",
+                         label = "Group Variable 2",
+                         choices = cb_options,
+                         selected = "")
+      updateRadioButtons(session, "gvBox",
+                         label = "Group Variable",
+                         choices = cb_options,
+                         selected = "")
+      updateRadioButtons(session, "qBox",
+                         label = "Quantity",
+                         choices = cb_options,
+                         selected = "")
+      updateRadioButtons(session, "gvMosaic1",
+                         label = "Group Variable 1",
+                         choices = cb_options,
+                         selected = "")
+      updateRadioButtons(session, "gvMosaic2",
+                         label = "Group Variable 2",
+                         choices = cb_options,
+                         selected = "")
+      updateRadioButtons(session, "gvScatter",
+                         label = "Group Variable",
+                         choices = cb_options,
+                         selected = "")
[…]
