+++ shiny-app/webinar-bio152/ui.R
@@ -1,0 +1,441 @@
+library(shiny)
+library(plotly)
+library(shinydashboard)
+
+shinyUI(dashboardPage(
+  skin="red",
+  dashboardHeader(
+    
+    title= tags$a(href='#',
+                  tags$img(src='fast-stats-logo-yellow-1.png',width=210,height =50)),
+    tags$li(img(src='red.png',width=10,height =50),class="dropdown"),
+    tags$li(img(src='logo.png',width=175,height =50),class="dropdown"),
+    tags$li(img(src='red.png',width=10,height =50),class="dropdown"),
+    tags$li(img(src='WIDSimpleAcronym.png',width=80,height =50),class="dropdown"),
+    tags$li(img(src='red.png',width=10,height =50),class="dropdown"),
+    tags$li(img(src='uw-logo-flush-web.png',width=150,height =50),class="dropdown"),
+    tags$li(img(src='red.png',width=10,height =50),class="dropdown")
+  ),
+  
+  dashboardSidebar(
+    sidebarMenu(
+      menuItem("About", tabName = "about"),
+      menuItem("Data Upload", tabName = "dataUpload"),
+      menuItem("Data Visualization", tabName = "dataVisualization"),
+      menuItem("Data Analysis", tabName = "dataAnalysis"),
+      menuItem("FAQ", tabName = "FAQ")
+    )
+  ),
+  
+  dashboardBody(
+    # for internal links
+    tags$script(HTML("
+        var openTab = function(tabName){
+          $('a', $('.sidebar')).each(function() {
+            if(this.getAttribute('data-value') == tabName) {
+              this.click()
+            };
+          });
+        }
+      ")),
+    tags$head(
+      tags$style(HTML("
+      .shiny-output-error-validation {
+        color: red;
+        font-size:120%;
+      },
+    ")),
+      # font for shiny notification
+      tags$style(
+        HTML(".shiny-notification {
+             position:fixed;
+             top: calc(40%);
+             left: calc(40%);
+             font-size:120%;
+             }
+             "
+        )
+      )
+    ),
+    
+    tabItems(
+      tabItem("about",
+              box( width=12, status="danger",solidHeader = T,
+                   title = "Welcome to the WI Fast Stats app: Cotelydon!",
+                   
+                   h4(HTML( 'WI Fast Stats is the <a href="https://github.com/crsl4/fast-stats" target="_blank">open-source</a> publicly available web app to analyze data from <a href="https://fastplants.org/" target="_blank">WI Fast Plants</a>.')),
+
+                   h4(HTML( 'This web app is the accompanying tool for the WI Fast Plants webinar: <a href="https://fastplants.org/2020/08/06/new-fast-plants-polycots-selection/" target="_blank"><i>Strategies for adapting WI Fast Plants selection of traits investigations for remote and social distance learning</i></a>.')),
+                   
+                   h4(HTML( 'A video of the Data Analysis part can be found <a href="https://www.currikistudio.org/playlist/6155/activity/167241/preview/lti" target="_blank">here</a>.'))
+                   
+                   ),
+              
+             
+              
+
+
+                
+      ),
+      
+      tabItem("dataUpload",
+              fluidRow(
+                box(
+                  width = 8, status = "danger",solidHeader = T,
+                  title = "File Display",
+                  tableOutput("contents"),
+                ),
+                box(
+                  width = 4, status = "danger", solidHeader = TRUE,
+                  title = "Data Upload",
+                  # Input: Select a file ----
+                  
+                  radioButtons("fileType","Use Sample Data or Upload Data",
+                               choices = c("Sample File" = "sampleFile",
+                                           "Upload File" = "uploadFile"),
+                               selected = "uploadFile"),
+                  
+                  conditionalPanel(condition="input.fileType=='uploadFile'", 
[…]
