+++ shiny-app/webinar-bio152/ui.R
@@ -70,12 +70,6 @@ shinyUI(dashboardPage(
                    h4(HTML( 'A video of the Data Analysis part can be found <a href="https://www.currikistudio.org/playlist/6155/activity/167241/preview/lti" target="_blank">here</a>.'))
                    
                    ),
-              
-             
-              
-
-
-                
       ),
       
       tabItem("dataUpload",
@@ -347,53 +341,51 @@ shinyUI(dashboardPage(
             # successfully hide the annoying verbaltext output using conditionalpanel
             # surprised to find that input$actionbuttonID is actually integer, keep adding one as users click the buttons
             conditionalPanel(condition="input.testType==0&&input.goChi!=0",
-                             verbatimTextOutput("chitest")),
+                             verbatimTextOutput("anovatest")),
             conditionalPanel(condition="input.testType==1&&input.goT!=0",
                              verbatimTextOutput("ttest"))
           ),
           box(
             width = 4, status = "danger",solidHeader = T,
             title = "Data Analysis",
-            
-            
-            selectInput("testType","Test Type",choices =c("Chi-Square Test"=0,"T-Test"=1)),
+            selectInput("testType","Test Type",choices =c("Anova Test"=0,"Summarise"=1)),
             # t-test
             conditionalPanel(condition="input.testType==1",
-                             h4("Data Analysis Option:"),
+                             # h4("Data Analysis Option:"),
                              
-                             checkboxInput("isEqualVar", "Equal Variance", F),
-                             
-                             HTML('<p style="color:#808080"> <b> Equal variance:</b> The standard t test assumes equal variances on the two groups. If the user checks this option, the standard t test is run, but if the user unchecks this option, then the Welch t test is run instead (that does not assume equal variances). </p>'),
+                             # checkboxInput("isEqualVar", "Equal Variance", F),
+                             # 
+                             # HTML('<p style="color:#808080"> <b> Equal variance:</b> The standard t test assumes equal variances on the two groups. If the user checks this option, the standard t test is run, but if the user unchecks this option, then the Welch t test is run instead (that does not assume equal variances). </p>'),
                              fluidRow(
                                column(6,radioButtons("groupVar","Group Variable:", c("1"="UnSpecified_Value","2"="UnSpecified_Value"))),
                                column(6,radioButtons("quantity","Quantity:", c("1"="UnSpecified_Value","2"="UnSpecified_Value")))
                              ),
-                             uiOutput("sel1"),
-                             uiOutput("sel2"),
-                             HTML('<p style="color:#808080"> <b>T test:</b> Statistical test of the null hypothesis of equality of means of a numerical variable ("Quantity") on two groups ("Group variable"). If the selected group variable has more than two categories, the user will select the two groups to compare. </p>'),
+                             # uiOutput("sel1"),
+                             # uiOutput("sel2"),
+                             # HTML('<p style="color:#808080"> <b>T test:</b> Statistical test of the null hypothesis of equality of means of a numerical variable ("Quantity") on two groups ("Group variable"). If the selected group variable has more than two categories, the user will select the two groups to compare. </p>'),
                              fluidRow(
                                column(6, align="center", offset = 3,
-                                      actionButton("goT", "T Test"),
+                                      actionButton("goT", "Summarise"),
                                       tags$style(type='text/css', "#button { vertical-align: middle; height: 50px; width: 100%; font-size: 30px;}")
                                )
                              ),
-                             HTML('<p style="color:#808080"> <b>How to interpret the result?</b> If the p-value is less than 0.05, we reject the null hypothesis of equality of means. The confidence interval represents the interval for the difference of means. </p>'),
[…]
