+++ shiny-app/webinar-bio152/server.R
@@ -9,6 +9,7 @@ library(thatssorandom)
 library(rsconnect)
 library(viridis)
 library(RColorBrewer)
+library(dplyr)
 # library(shinyjs)
 # Define server logic to read selected file ----
 # global variable
@@ -62,11 +63,11 @@ server <- function(input, output,session) {
                          choices = cb_options,
                          selected = "")
       updateRadioButtons(session, "gv1",
-                         label = "Group Variable 1",
+                         label = "Group Variable",
                          choices = cb_options,
                          selected = "")
       updateRadioButtons(session, "gv2",
-                         label = "Group Variable 2",
+                         label = "Quantity Variable",
                          choices = cb_options,
                          selected = "")
       updateRadioButtons(session, "gvBox",
@@ -119,19 +120,19 @@ server <- function(input, output,session) {
                          choices = c("plant.ID"="plant.ID","cotyledons"="cotyledons","generation"="generation","parents"="parents"),
                          selected = "")
       updateRadioButtons(session, "gv1",
-                         label = "Group Variable 1",
-                         choices = c("plant.ID"="plant.ID","cotyledons"="cotyledons","generation"="generation","parents"="parents"),
-                         selected = "")
-      updateRadioButtons(session, "gv2",
-                         label = "Group Variable 2",
-                         choices = c("plant.ID"="plant.ID","cotyledons"="cotyledons","generation"="generation","parents"="parents"),
-                         selected = "")
-      updateRadioButtons(session, "gvBox",
                          label = "Group Variable",
                          choices = c("plant.ID"="plant.ID","cotyledons"="cotyledons","generation"="generation","parents"="parents"),
                          selected = "")
-      updateRadioButtons(session, "qBox",
+      updateRadioButtons(session, "gv2",
                          label = "Quantity",
+                         choices = c("plant.ID"="plant.ID","cotyledons"="cotyledons","generation"="generation","parents"="parents"),
+                         selected = "")
+      updateRadioButtons(session, "gvBox",
+                         label = "Group Variable 1",
+                         choices = c("plant.ID"="plant.ID","cotyledons"="cotyledons","generation"="generation","parents"="parents"),
+                         selected = "")
+      updateRadioButtons(session, "qBox",
+                         label = "Group Variable 2",
                          choices = c("plant.ID"="plant.ID","cotyledons"="cotyledons","generation"="generation","parents"="parents"),
                          selected = "")
       updateRadioButtons(session, "gvMosaic1",
@@ -1009,163 +1010,34 @@ server <- function(input, output,session) {
     validate(
       not_quantity(df,v8$q)
     )
-    group_val<-unlist(subset(df, select=c(v7$gv)))
+    
+    group_var<-unlist(subset(df, select=c(v7$gv)))
+    q_var<-unlist(subset(df, select=c(v8$q)))
+    
+    
+    group_by(df,v7$gv) %>%
+      summarise(
+        count = n(),
+        mean = mean(q_var, na.rm = TRUE),
+        sd = sd(q_var, na.rm = TRUE)
+      )
     
     # , in the end omits default val set to be True
     # # important
-    x = subset(df, select=c(v8$q))[group_val == v14$sel1,]
[…]
