### 2021-04-01 78f9e9e
solve bugs in summarize and lm
