### 2021-04-22 #103 @crsl4
Hello @labarba ,
No pressure at all, but just wondering if you have any updates on the JOSE scope and the suitability of our work.
Again, just touching base and as a fellow reviewer/editor, I completely understand the speed of review these days :)
