### 2021-04-26 b7b4c2b
finished implementing all the warnings
