+++ shiny-app/webinar-bio152/server.R
@@ -194,6 +194,8 @@ server <- function(input, output,session) {
       df<-data_set()
     }
     validate(not_empty(df))
+    validate(not_exp(df))
+    validate(not_big_dataset(df))
     v$doViolinPlot <- input$goViolin
     
   })
@@ -225,6 +227,8 @@ server <- function(input, output,session) {
       df<-data_set()
     }
     validate(not_empty(df))
+    validate(not_exp(df))
+    validate(not_big_dataset(df))
     # 0 will be coerced to FALSE
     # 1+ will be coerced to TRUE
     v4$doT <- input$goT
@@ -262,6 +266,8 @@ server <- function(input, output,session) {
       df<-data_set()
     }
     validate(not_empty(df))
+    validate(not_exp(df))
+    validate(not_big_dataset(df))
     # 0 will be coerced to FALSE
     # 1+ will be coerced to TRUE
     v9$doScatter <- input$goScatter
@@ -280,6 +286,8 @@ server <- function(input, output,session) {
       df<-data_set()
     }
     validate(not_empty(df))
+    validate(not_exp(df))
+    validate(not_big_dataset(df))
     v10$doBox <- input$goBox
   })
   
@@ -373,6 +381,8 @@ server <- function(input, output,session) {
       df<-data_set()
     }
     validate(not_empty(df))
+    validate(not_exp(df))
+    validate(not_big_dataset(df))
     # 0 will be coerced to FALSE
     # 1+ will be coerced to TRUE
     v22$doChi <- input$goChi
@@ -406,6 +416,8 @@ server <- function(input, output,session) {
       df<-data_set()
     }
     validate(not_empty(df))
+    validate(not_exp(df))
+    validate(not_big_dataset(df))
     v27$doMosaic <- input$goMosaic
   })
   v28 <- reactiveValues(doAddPoints2 =T)
@@ -584,6 +596,20 @@ server <- function(input, output,session) {
     }
   }
   
+  not_exp<-function(df){
+    if("Experiment" %in% colnames(df)){
+      
+    }else{
+      showNotification("Please upload a file that contains \"Experiment\" column!", duration=3, type="error")
+    }
+  }
+  
+  not_big_dataset<-function(df){
+    if(nrow(df)<15){
+      showNotification("Please make sure that the Experiment you've selected contain at least 15 rows!", duration=3, type="error")
+    }
+  }
+  
   output$summary <- renderPrint({
     
     # input$file1 will be NULL initially. After the user selects
@@ -616,11 +642,14 @@ server <- function(input, output,session) {
     else
     {
       df<-data_set()
+      # not_exp(df)
+      # choice<-v54$choice1
+      # df<-df[df$Experiment==choice,c(1:ncol(df))]
     }
     # choice<-v54$sec1
     # colName<-v53$section
     # df1<-subset(df, colName==choice)
-   
+    # validate(not_exp(df))
     if(input$disp == "head") {
       return(head(df))
     }
@@ -646,9 +675,13 @@ server <- function(input, output,session) {
     }
     else
     {
-      df<-data_set()
+      df<-data_set
+      validate(not_exp(df))
+      choice<-v54$choice1
+      df<-df[df$Experiment==choice, c(1:ncol(df))]
     }
     req(df)
+    validate(not_big_dataset(df))
    
     # print(df)
     if (v$doViolinPlot == FALSE) return()
@@ -696,6 +729,7 @@ server <- function(input, output,session) {
     # and uploads a file, head of that data file by default,
     # or all rows if selected, will be shown.
     
+    
     if(input$fileType=="sampleFile")
     {
       df<- toy
[…]
