### 2021-05-06 c8b0195
remove inf
