+++ shiny-app/webinar-bio152/ui.R
@@ -61,13 +61,15 @@ shinyUI(dashboardPage(
     tabItems(
       tabItem("about",
               box( width=12, status="danger",solidHeader = T,
-                   title = "Welcome to the WI Fast Stats app: Cotelydon!",
+                   title = "Welcome to the WI Fast Stats web-app for IntroBio!",
                    
                    h4(HTML( 'WI Fast Stats is the <a href="https://github.com/crsl4/fast-stats" target="_blank">open-source</a> publicly available web app to analyze data from <a href="https://fastplants.org/" target="_blank">WI Fast Plants</a>.')),
 
-                   h4(HTML( 'This web app is the accompanying tool for the WI Fast Plants webinar: <a href="https://fastplants.org/2020/08/06/new-fast-plants-polycots-selection/" target="_blank"><i>Strategies for adapting WI Fast Plants selection of traits investigations for remote and social distance learning</i></a>.')),
-                   
-                   h4(HTML( 'A video of the Data Analysis part can be found <a href="https://www.currikistudio.org/playlist/6155/activity/167241/preview/lti" target="_blank">here</a>.'))
+                   h4(HTML( 'This web app is the accompanying tool for the UW-Madison class: <a href="https://introbio.integrativebiology.wisc.edu/" target="_blank">Introductory Biology 151/153-152</a>.')),
+
+                   h4(HTML( 'A 9-minute tutorial of the WI Fast Stats web app can be found <a href="https://www.youtube.com/watch?v=6vX3mVXcK4s&t=4s" target="_blank">here</a>.')),
+                                      
+                   h4(HTML( 'For more information, a video of the Data Analysis part of a WI Fast Plants webinar can be found <a href="https://www.currikistudio.org/playlist/6155/activity/167241/preview/lti" target="_blank">here</a>.'))
                    
                    ),
       ),
@@ -91,15 +93,15 @@ shinyUI(dashboardPage(
                   
                   conditionalPanel(condition="input.fileType=='uploadFile'", 
                                    
-                                   HTML('<p style="color:#808080">Data in delimited text files can be separated by comma, tab or semicolon. For example, Excel data can be exported in .csv (comma separated) or .tab (tab separated) format </p>'),
+                                   HTML('<p style="color:#808080">Data in delimited text files can be separated by comma, tab or semicolon. For example, Excel data can be exported in .csv (comma separated) or .tab (tab separated) format. A template of the spreadsheet can be found <a href="https://docs.google.com/spreadsheets/d/12fEIaSCUdT8s16jPEZ3grV6pQeWsy6pcHtwuRlbW5JM/edit?usp=sharing" target="_blank">here</a> </p>'),
                                    
                                    fileInput("file1", "Choose CSV File or Drag the file here",
                                              multiple = FALSE,
                                              accept = c("text/csv",
                                                         "text/comma-separated-values,text/plain",
                                                         ".csv")),
-                                   HTML('<p style="color:#808080"> <b>Notice:</b> The maximum file size should not exceed <b>10MB</b></p>'),
-                                   HTML('<p style="color:#808080"> <b>Notice:</b> The uploaded file must have a column named<b> Experiment</b></p>'),
+                                   HTML('<p style="color:#808080"> <b>Warning:</b> The maximum file size should not exceed <b>10MB</b></p>'),
+                                   HTML('<p style="color:#808080"> <b>Warning:</b> The uploaded file must have a column named<b> Experiment</b>. A template of the spreadsheet can be found <a href="https://docs.google.com/spreadsheets/d/12fEIaSCUdT8s16jPEZ3grV6pQeWsy6pcHtwuRlbW5JM/edit?usp=sharing" target="_blank">here</a></p>'),
                                    
                                    # Input: Checkbox if file has header ----
                                    tags$p("Click the checkbox if file has a header:"),
[…]
