+++ shiny-app/web-app-recipe.md
@@ -10,6 +10,21 @@ If you are interested to add a new web-app to WI Fast Stats, these are some step
 
 4. If you want to change the description on the About page, you do this in `ui.R`, lines 66-70
 
+# How to add to the website
+
+1. Navigate to Fast-stats/website/index.html
+
+2. Navigate to https://wi-fast-stats.wid.wisc.edu/, and right click the `WEBAPPS` section with `Inspect` to see where the code is
+
+3. Back to index.html, please follow the style given and add the web-app link to a new block
+
+# How to add a DOCS section
+
+1. Navigate to fast-stats/DOCS.md
+
+2. You can then modify the content here
+
+
 missing steps to include:
 - how to add to the website
 - how to deploy a specific web-app (I think that this needs to be done by WID people)
