+++ shiny-app/web-app-recipe.md
@@ -1,34 +1,39 @@
 # How to add a new web-app to WI Fast Stats
 
-If you are interested to add a new web-app to WI Fast Stats, these are some steps that you can follow. Make sure to always work on `development` branch of your fork.
+This is an internal document for the [Solis-Lemus lab](https://solislemuslab.github.io/). Outside contributors who would like to create new WI Fast Stats web apps, please contact Claudia Solis-Lemus (contact info in the [website](https://solislemuslab.github.io//pages/people.html)) and make sure to read the [contributing doc](https://github.com/crsl4/fast-stats/blob/master/CONTRIBUTING.md).
 
-1. You can copy the existing folder from one of the web apps e.g. `webinar-aug20`. You can choose the web-app that has some of the features that you need
+For contributors within the Solis-Lemus lab: if you are interested to add a new web-app to WI Fast Stats, these are some steps that you can follow. Make sure to always work on `development` branch of your fork.
 
-2. You need to change the `ui.R` which is what the users see and `server.R` which is the internal code that runs the plots and data analyses.
-    - Make sure to read the official `R shiny` and `shiny dashboard` documentation before making any changes [add links]
+1. You can copy the existing folder from one of the web apps e.g. `webinar-aug20` since these folders already have the right structure and web app format. You can choose the web-app that has some of the features that you need (e.g. in terms of Data Visualization or Data Analysis)
 
-3. If you want to change the color of the web-app, you do that in `ui.R`. For example, in the `webinar-aug20` folder, in line 6 and also in line 63: the variable `status` specifies the color of the web-app. For example, `status="danger"` is a red color
+2. You need to change the `ui.R` file which is what the users see and `server.R` file which is the internal code that runs the plots and data analyses.
+    - Make sure to read the official `R shiny` and `shiny dashboard` documentation before making any changes
 
-4. If you want to change the description on the About page, you do this in `ui.R`, lines 66-70 in the `webinar-aug20` example
+3. The `ui.R` file controls what the users see. We refer to [`ui.R` file in the `webinar-aug20` example](https://github.com/crsl4/fast-stats/blob/master/shiny-app/webinar-aug20/ui.R) for the following lines of code:
+  - If you want to change the color of the web-app, you do that in `ui.R`. For example, in the `ui.R` file in the `webinar-aug20` folder, in line 6 and also in line 63: the variable `status` specifies the color of the web-app. For example, `status="danger"` is a red color
+  - If you want to change the description on the About page, you do this in lines 66-70 in the `webinar-aug20` example
+  - To change the the right-side menu, you can check lines 20-28
+  - The Data Upload part starts on line 81
+  - The Data Visualization part starts on line 135
+  - The Data Analysis part starts on line 341
 
-5. Lines 20-28 in the `webinar-aug20` example contain the right-side menus
-
-6. The Data Upload part starts on line 81 in the `webinar-aug20` example 
-
-7. The Data Visualization part starts on line 135 in the `webinar-aug20` 
-example
-
-8. The Data Analysis part starts on line 341 in the `webinar-aug20` example
+9. The `server.R` file contains the actual code that produces the plots and analyses. We refer to [`server.R` file in the `webinar-aug20` example](https://github.com/crsl4/fast-stats/blob/master/shiny-app/webinar-aug20/server.R) for the following lines of code:
+  - Starting on line 42, you have the function to upload a file with a choice between "upload file" or using the sample data already included in the web app
+  - All the reactive values correspond to choices made by the user in the left-side menu. For example, in line 260 we have the `doBox` option of whether to do the box plot. This selection (binary=true/false) is stored in the `v10$doBox` variable
[…]
