### 2021-11-05 #103 @crsl4
I had this error too, but it does compile locally per the instructions. Not sure what the error is about!
