### 2021-11-18 #103 @xuanxu
The PDF is not compiling due to an unscaped character. [This change](https://github.com/crsl4/fast-stats/pull/51) should fix it.
