### 2021-12-03 #103 @JasonJWilliamsNY
> The PDF is not compiling due to an unscaped character. [This change](https://github.com/crsl4/fast-stats/pull/51) should fix it.

Thanks @xuanxu @crsl4 can you accept the change
