### 2022-02-14 #159 @whedon
[REVIEW]: WI Fast Stats: a collection of web apps for the visualization and analysis of WI Fast Plants data
**Submitting author:** @crsl4 (<a href="http://orcid.org/0000-0002-9789-8915">Claudia Solis-Lemus</a>)
**Repository:** <a href="https://github.com/crsl4/fast-stats" target ="_blank">https://github.com/crsl4/fast-stats</a>
**Version:** v2.1

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode".

- [x] #160
