### 2022-02-24 crsl4/fast-stats#52 @pierrerogy
JOSS Review 1
Hello!

First, great job on the apps, they look great and I am sure they will indeed be a fundamental step in teaching data science. I found the documentation thorough and clear, and the apps engaging and fun to use. Please find below my comments associated with the checklist in the JOSS review. Please let me know if you have any questions!

### General checks
Everything looks good!

### Functionality & Documentation
I was able to download the repo successfully using the provided command. However, I had several issues with the rest of the provided code.

-  The package `thatssorandom` is not available on at least the two most recent versions of R, so I had to use `install.github(“https://github.com/EdwinTh/thatssorandom")` to get it and use the apps
- I could not run the app using the provided command (`shiny::runApp("/path/to/wi-fast-stats")`). I think the first issue here is that the repo downloads as `fast-stats` and not `wi-fast-stats`. Then, the provided command should directly point to the folder where one of the app is present in `shiny-app/`.
- I think the code could use a bit more commenting. It took me some time to wrap my head around what was happening. More importantly, as this app is intended to a younger audience, it could be their entry door to coding. If the code is adequately commented, it will definitely be more understandable by students (and teachers/researchers!) and thus have broader impact.

In terms of the app functionnality, I only have two main comments (only using sample data):

- I think use of the app would be made a lot easier if the "Group variable" and "Quantity" radiobuttons could only show the corresponding variables, instead of all the columns (categorical and numerical variables, respectively).
- It may be worth it to let plant ID be a "Group variable"? I am not sure if yes, but my data mind would love to see a raw data plot or a histogram.

### Software paper
- Williams and Hill (1986) is missing DOI: [10.1126/science.232.4756.1385](https://doi.org/10.1126/science.232.4756.1385)
- The paper refers to a section called "Data summary" but the apps have either nothing or "Data upload", I would make section names constistent.

Best of luck on the rest, and can't wait to see this project completed!

Pierre
