### 2022-02-24 #159 @JasonJWilliamsNY
Thanks @pierrerogy I see your issue/comments in the project repository (https://github.com/crsl4/fast-stats/issues/52) they look detailed and helpful. Also thanks for the Checklist. @crsl4 will also find them useful I think. @magitz hope you will have a chance to looks soon.
