### 2022-02-26 crsl4/fast-stats#52 @yizhou0522
Hi @pierrerogy @xuanxu ,

I'm Yizhou, one of the authors of this app. Thank you so much for the constructive comments on our work. I'll use them to improve our work and submit a more completed version later. By the way, are there any deadlines for us to resubmit our revised work?
