### 2022-02-26 crsl4/fast-stats#52 @crsl4
Echoing Yizhou, thanks so much for the constructive feedback!
@JasonJWilliamsNY should we wait to receive both reviews to make changes or should we address them as they come?
Thanks!
