### 2022-03-07 crsl4/fast-stats#53 @magitz
JOSS review 2
Overall this looks like a handy contribution. The apps provide a user-friendly way to interact with data for a fairly defined project.

I have a few suggestions below.

## Authorship

- [x] Yizhou Liu seems to have contributed one of the slide decks and a fair bit of the commits. I wonder if they should be included in the authorship? Or acknowledged more in the paper?

## Functionality & Documentation

- [x] As noted in #52 , `thatssorandom` couldn't be installed. I was able to get it working with that issue's guidance.
- [x] Overall, I think [this section](https://github.com/crsl4/fast-stats#running-the-web-apps-locally) needs work. Starting the application as indicated does not work.
- [x] When running the `webinar-aug20` app to make a MosaicPlot, I got the error:
   ```R
   Warning: `guides(<scale> = FALSE)` is deprecated. Please use `guides(<scale> = "none")` instead.
   ```
- [x] I think expanding documentation would help facilitate using this tool in teaching. For example, the error messages for making plots can be a bit frustrating--maybe identify which columns of the dataset are categorical vs numerical? Or provide more description of the sample dataset? Or limit selections to the appropriate variable type?
- [x] I like that the slides are part of the github repo, would it be possible to share the Google Slides link as well?
   - [x] Would it be possible to share the datafiles used in making the figures in the slides? Or provide a copy that matches the sample file? In going through the slides, I tried to re-create the graphs and while similar, that are different. I assume different data were used, but would be nice to be able to replicate the examples provided.
   - [ ] I think this could also serve as the start of adding tests. I didn't investigate the code for the Chi-square, etc, Should probably add tests.
- [x]  `Warning: Groups with fewer than two data points have been dropped.` is reported in R for some of the plots, but no feedback is given to the App. Maybe that should be passed through?

### `notebooks` folder
Not really sure what this folder is supposed to be. The notebooks do not work as-is. e.g. there is no `data/ecosystem-data.csv` file to load as on line 18 of [ecosystem-data-analysis.Rmd](https://github.com/crsl4/fast-stats/blob/master/notebooks/ecosystem-data-analysis.Rmd).

- [x] **References**: Please add DOIs to all papers.

## Editorial

- [x] [DOCS.md](https://github.com/crsl4/fast-stats/blob/master/DOCS.md): "The Home page is a collection of animation using WI Fast Plants photos.": "collection of animation" seems like a poor word choice here.
- [ ] Webinar links in timeline seem broken. Not necessarily something that needs fixing, but might want to either fix or remove links.
  - [ ] https://www.carolina.com/knowledge/2020/10/20/webinar-student-designed-ecosystem-experiments-in-middle-grades - 404 Error
  - [ ] https://fastplants.org/2020/08/06/new-fast-plants-polycots-selection/ - Redirect to main site
