### 2022-03-07 #159 @magitz
I've created https://github.com/crsl4/fast-stats/issues/53
Ignore the question on authorship...I didn't see the pdf was available above until just now (it's not in the ms folder for some reason) and the table on the paper.md file is a bit of a pain to read and I didn't see that Yizhou is indeed included when scrolling through it.

One other thing that catches my attention now is I am not sure how useful this would be for elementary and even middle school.
