+++ shiny-app/webinar-aug20/server.R
@@ -45,6 +45,7 @@ server <- function(input, output,session) {
       dsnames <- names(data_set())
       cb_options <- list()
       cb_options[ dsnames] <- dsnames
+      print(cb_options)
       updateRadioButtons(session, "xaxisGrp",
                          label = "Group Variable",
                          choices = cb_options,
@@ -102,61 +103,64 @@ server <- function(input, output,session) {
                          choices = cb_options,
                          selected = "")
     }else{
+      # "sample data" option
+      gv_vec <- c("plant.ID"="plant.ID","generation"="generation","parents"="parents")
+      q_vec <- c("cotyledons"="cotyledons")
       updateRadioButtons(session, "xaxisGrp",
                          label = "Group Variable",
-                         choices = c("plant.ID"="plant.ID","cotyledons"="cotyledons","generation"="generation","parents"="parents"),
+                         choices =  gv_vec,
                          selected = "")
       updateRadioButtons(session, "yaxisGrp",
                          label = "Quantity",
-                         choices = c("plant.ID"="plant.ID","cotyledons"="cotyledons","generation"="generation","parents"="parents"),
+                         choices = q_vec,
                          selected = "")
       updateRadioButtons(session, "groupVar",
                          label = "Group Variable",
-                         choices = c("plant.ID"="plant.ID","cotyledons"="cotyledons","generation"="generation","parents"="parents"),
+                         choices =  gv_vec,
                          selected = "")
       updateRadioButtons(session, "quantity",
                          label = "Quantity",
-                         choices = c("plant.ID"="plant.ID","cotyledons"="cotyledons","generation"="generation","parents"="parents"),
+                         choices =q_vec,
                          selected = "")
       updateRadioButtons(session, "gv1",
                          label = "Group Variable 1",
-                         choices = c("plant.ID"="plant.ID","cotyledons"="cotyledons","generation"="generation","parents"="parents"),
+                         choices =  gv_vec,
                          selected = "")
       updateRadioButtons(session, "gv2",
                          label = "Group Variable 2",
-                         choices = c("plant.ID"="plant.ID","cotyledons"="cotyledons","generation"="generation","parents"="parents"),
+                         choices =  gv_vec,
                          selected = "")
       updateRadioButtons(session, "gvBox",
                          label = "Group Variable",
-                         choices = c("plant.ID"="plant.ID","cotyledons"="cotyledons","generation"="generation","parents"="parents"),
+                         choices = gv_vec,
                          selected = "")
       updateRadioButtons(session, "qBox",
                          label = "Quantity",
-                         choices = c("plant.ID"="plant.ID","cotyledons"="cotyledons","generation"="generation","parents"="parents"),
+                         choices =q_vec,
                          selected = "")
       updateRadioButtons(session, "gvMosaic1",
                          label = "Group Variable 1",
-                         choices = c("plant.ID"="plant.ID","cotyledons"="cotyledons","generation"="generation","parents"="parents"),
+                         choices =  gv_vec,
                          selected = "")
       updateRadioButtons(session, "gvMosaic2",
[…]
