### 2022-03-12 8a30810
complete only show gv and q options feature
