+++ shiny-app/webinar-aug20/server.R
@@ -13,6 +13,9 @@ library(RColorBrewer)
 # Define server logic to read selected file ----
 # global variable
 options(shiny.maxRequestSize=10*1024^2)
+# ---------suppress all R's warning-------------
+# options(warn=-1)
+# --------------------------------------------
 toy<-read.csv("toys.csv",
               header = T,
               sep = ",")
@@ -42,70 +45,91 @@ server <- function(input, output,session) {
   outputOptions(output, 'fileUploaded', suspendWhenHidden=FALSE)
   observe({
     if(input$fileType=="uploadFile"){
+      df <- data_set()
       dsnames <- names(data_set())
-      cb_options <- list()
-      cb_options[ dsnames] <- dsnames
-      print(cb_options)
+      cat("dsnames:  ", typeof(dsnames), dsnames)
+      print("\n")
+      gv_options <- list()
+      q_options <- list()
+      for(var in dsnames){
+        # cat("xxx", typeof(var), var)
+        print("\n")
+        if(is_quantity(df, var)){
+          q_options[var] <- var
+        }else{
+          gv_options[var] <- var
+        }
+      }
+      # print(q_options)
+      # print("\n")
+      # print(">>>>>>>>>>>>>>>")
+      # print(gv_options)
+      # print("\n")
+      # cb_options <- list()
+      # 
+      # cb_options[ dsnames] <- dsnames
+      # cat("cb_options  ",cb_options)
       updateRadioButtons(session, "xaxisGrp",
                          label = "Group Variable",
-                         choices = cb_options,
+                         choices = gv_options,
                          selected = "")
       updateRadioButtons(session, "yaxisGrp",
                          label = "Quantity",
-                         choices = cb_options,
+                         choices = q_options,
                          selected = "")
       updateRadioButtons(session, "groupVar",
                          label = "Group Variable",
-                         choices = cb_options,
+                         choices = gv_options,
                          selected = "")
       updateRadioButtons(session, "quantity",
                          label = "Quantity",
-                         choices = cb_options,
+                         choices = q_options,
                          selected = "")
       updateRadioButtons(session, "gv1",
                          label = "Group Variable 1",
-                         choices = cb_options,
+                         choices = gv_options,
                          selected = "")
       updateRadioButtons(session, "gv2",
                          label = "Group Variable 2",
-                         choices = cb_options,
+                         choices = gv_options,
                          selected = "")
       updateRadioButtons(session, "gvBox",
                          label = "Group Variable",
-                         choices = cb_options,
+                         choices = gv_options,
                          selected = "")
       updateRadioButtons(session, "qBox",
                          label = "Quantity",
-                         choices = cb_options,
+                         choices = q_options,
                          selected = "")
       updateRadioButtons(session, "gvMosaic1",
                          label = "Group Variable 1",
-                         choices = cb_options,
+                         choices = gv_options,
                          selected = "")
       updateRadioButtons(session, "gvMosaic2",
                          label = "Group Variable 2",
-                         choices = cb_options,
+                         choices = gv_options,
                          selected = "")
       updateRadioButtons(session, "gvScatter",
                          label = "Group Variable",
-                         choices = cb_options,
+                         choices = gv_options,
                          selected = "")
       updateRadioButtons(session, "qScatter",
                          label = "Quantity",
-                         choices = cb_options,
[…]
