+++ shiny-app/webinar-dec20/server.R
@@ -39,123 +39,117 @@ server <- function(input, output,session) {
   })
   observe({
     if(input$fileType=="uploadFile"){
+      df <- data_set()
       dsnames <- names(data_set())
-      cb_options <- list()
-      cb_options[ dsnames] <- dsnames
+      gv_options <- list()
+      q_options <- list()
+      for(var in dsnames){
+        if(is_quantity(df, var)){
+          q_options[var] <- var
+        }else{
+          gv_options[var] <- var
+        }
+      }
       updateRadioButtons(session, "xaxisGrp",
                          label = "Group Variable",
-                         choices = cb_options,
+                         choices = gv_options,
                          selected = "")
       updateRadioButtons(session, "yaxisGrp",
                          label = "Quantity",
-                         choices = cb_options,
+                         choices = q_options,
                          selected = "")
       updateRadioButtons(session, "groupVar",
                          label = "Group Variable",
-                         choices = cb_options,
+                         choices = gv_options,
                          selected = "")
       updateRadioButtons(session, "quantity",
                          label = "Quantity",
-                         choices = cb_options,
+                         choices = q_options,
                          selected = "")
       updateRadioButtons(session, "gv1",
                          label = "Group Variable 1",
-                         choices = cb_options,
+                         choices = gv_options,
                          selected = "")
       updateRadioButtons(session, "gv2",
                          label = "Group Variable 2",
-                         choices = cb_options,
+                         choices = gv_options,
                          selected = "")
       updateRadioButtons(session, "gvBox",
                          label = "Group Variable",
-                         choices = cb_options,
+                         choices = gv_options,
                          selected = "")
       updateRadioButtons(session, "qBox",
                          label = "Quantity",
-                         choices = cb_options,
+                         choices = q_options,
                          selected = "")
       updateRadioButtons(session, "gvMosaic1",
                          label = "Group Variable 1",
-                         choices = cb_options,
+                         choices = gv_options,
                          selected = "")
       updateRadioButtons(session, "gvMosaic2",
                          label = "Group Variable 2",
-                         choices = cb_options,
+                         choices = gv_options,
                          selected = "")
       updateRadioButtons(session, "gvScatter",
                          label = "Group Variable",
-                         choices = cb_options,
+                         choices = gv_options,
                          selected = "")
       updateRadioButtons(session, "qScatter",
                          label = "Quantity",
-                         choices = cb_options,
+                         choices = q_options,
                          selected = "")
       updateRadioButtons(session, "gvDensities",
                          label = "Group Variable",
-                         choices = cb_options,
[…]
