+++ shiny-app/webinar-aug20/server.R
@@ -1,3 +1,25 @@
+#####################################################################
+#
+# server.R
+#
+# 
+#
+#     This program is free software; you can redistribute it and/or
+#     modify it under the terms of the GNU General Public License,
+#     version 3, as published by the Free Software Foundation.
+#
+#     This program is distributed in the hope that it will be useful,
+#     but without any warranty; without even the implied warranty of
+#     merchantability or fitness for a particular purpose.  See the GNU
+#     General Public License, version 3, for more details.
+#
+#     A copy of the GNU General Public License, version 3, is available
+#     at http://www.r-project.org/Licenses/GPL-3
+#
+# Part of the fast-stats package
+# Contains: server.R
+######################################################################
+
 library(data.table)
 library(ggplot2)
 library(car)
@@ -9,9 +31,6 @@ library(thatssorandom)
 library(rsconnect)
 library(viridis)
 library(RColorBrewer)
-# library(shinyjs)
-# Define server logic to read selected file ----
-# global variable
 options(shiny.maxRequestSize=10*1024^2)
 # ---------suppress all R's warning-------------
 # options(warn=-1)
@@ -19,8 +38,10 @@ options(shiny.maxRequestSize=10*1024^2)
 toy<-read.csv("toys.csv",
               header = T,
               sep = ",")
+######################################################################
+# R shiny server function
+######################################################################
 server <- function(input, output,session) {
-  # #################################### 
   dsnames<-c() #a vector to store col names
   
   data_set <- reactive({
@@ -43,6 +64,12 @@ server <- function(input, output,session) {
     return(!is.null(data_set()))
   })
   outputOptions(output, 'fileUploaded', suspendWhenHidden=FALSE)
+
+  ######################################################################
+  # update the radio buttons as user specified, i.e. uploadFile mode or sample
+  # file mode, then classify the options either as group variable or quantity
+  # variables.
+  ######################################################################  
   observe({
     if(input$fileType=="uploadFile"){
       df <- data_set()
@@ -176,16 +203,14 @@ server <- function(input, output,session) {
     }
     
   })
-  #####################testing
-  # observeEvent(input$fileType=="sampleFile", updateRadioButtons(session, "gv1",
-  #                                                      label = "Group Variable 1",
-  #                                                      choices = c("xxxx"="x"),
-  #                                                      selected = ""))
   
   output$choose_dataset <- renderUI({
     selectInput("dataset", "Data set", as.list(data_sets))
   })
-  ####################################
+  ######################################################################
+  # Following obserfverEvent functions
+  # used to assign local values from reactive values
+  ######################################################################
   v<-reactiveValues(doViolinPlot=FALSE)
   observeEvent(input$goViolin, {
     # 0 will be coerced to FALSE
@@ -520,7 +545,10 @@ server <- function(input, output,session) {
     v52$colorDensities<-input$colorDensities
   })
   
-  
+  ######################################################################
+  # not_equalGV: compare if two group variables are the same, show warning 
+  # message if they are the same
+  ######################################################################
   not_equalGV<-function(input1, input2){
     if(strcmp(input1, input2)){
       showNotification("Please select different 'Group Variables'!",duration=3,type = "error")
@@ -534,6 +562,10 @@ server <- function(input, output,session) {
     }
   }
   
+  ######################################################################
+  # not_equalGV2: compare if two group variables are the same, show warning 
[…]
