+++ shiny-app/webinar-aug20/ui.R
@@ -1,9 +1,39 @@
+#####################################################################
+#
+# ui.R
+#     
+# 
+#
+#     This program is free software; you can redistribute it and/or
+#     modify it under the terms of the GNU General Public License,
+#     version 3, as published by the Free Software Foundation.
+#
+#     This program is distributed in the hope that it will be useful,
+#     but without any warranty; without even the implied warranty of
+#     merchantability or fitness for a particular purpose.  See the GNU
+#     General Public License, version 3, for more details.
+#
+#     A copy of the GNU General Public License, version 3, is available
+#     at http://www.r-project.org/Licenses/GPL-3
+#
+# Part of the fast-stats package
+# Contains: ui.R
+######################################################################
+
+
 library(shiny)
 library(plotly)
 library(shinydashboard)
-
+######################################################################
+# 
+# shinyUI: Create a Shiny UI handler
+# ui:
+# A user interace definition
+# 
+###################################################################### 
 shinyUI(dashboardPage(
   skin="yellow",
+  # header
   dashboardHeader(
     
     title= tags$a(href='#',
@@ -17,6 +47,7 @@ shinyUI(dashboardPage(
     tags$li(img(src='yellow1.png',width=10,height =50),class="dropdown")
   ),
   
+  # sidebar 
   dashboardSidebar(
     sidebarMenu(
       menuItem("About", tabName = "about"),
@@ -59,6 +90,7 @@ shinyUI(dashboardPage(
     ),
     
     tabItems(
+      # about button
       tabItem("about",
               box( width=12, status="warning",solidHeader = T,
                    title = "Welcome to the WI Fast Stats app: Cotelydon!",
@@ -70,14 +102,9 @@ shinyUI(dashboardPage(
                    h4(HTML( 'A video of the Data Analysis part can be found <a href="https://www.currikistudio.org/playlist/6155/activity/167241/preview/lti" target="_blank">here</a>.'))
                    
                    ),
-              
-             
-              
-
-
-                
       ),
       
+      # data upload button
       tabItem("dataUpload",
               fluidRow(
                 box(
@@ -132,6 +159,7 @@ shinyUI(dashboardPage(
                 )
               )
       ),
+      # data visualization button
       tabItem(
         "dataVisualization",
         
