+++ shiny-app/webinar-dec20/server.R
@@ -1,3 +1,25 @@
+#####################################################################
+#
+# server.R
+#
+# 
+#
+#     This program is free software; you can redistribute it and/or
+#     modify it under the terms of the GNU General Public License,
+#     version 3, as published by the Free Software Foundation.
+#
+#     This program is distributed in the hope that it will be useful,
+#     but without any warranty; without even the implied warranty of
+#     merchantability or fitness for a particular purpose.  See the GNU
+#     General Public License, version 3, for more details.
+#
+#     A copy of the GNU General Public License, version 3, is available
+#     at http://www.r-project.org/Licenses/GPL-3
+#
+# Part of the fast-stats package
+# Contains: server.R
+######################################################################
+
 library(data.table)
 library(ggplot2)
 library(car)
@@ -13,10 +35,22 @@ options(shiny.maxRequestSize=10*1024^2)
 toy<-read.csv("2016-brassica.csv",
              header = T,
              sep = ",")
-
-# Define server logic to read selected file ----
+######################################################################
+# server: R shiny server function, 
+# 
+# input:
+# The session's input object (the same as is passed into the Shiny server 
+# function as an argument)
+# 
+# output:
+# The session's output object (the same as is passed into the Shiny 
+# server function as an argument)
+# 
+# shiny:
+# The session object is an environment that can be used to access 
+# information and functionality relating to the session. 
+######################################################################
 server <- function(input, output,session) {
-  # #################################### 
   dsnames<-c() #a vector to store col names
   
   data_set <- reactive({
@@ -37,6 +71,11 @@ server <- function(input, output,session) {
     
     
   })
+  ######################################################################
+  # update the radio buttons as user specified, i.e. uploadFile mode or sample
+  # file mode, then classify the options either as group variable or quantity
+  # variables.
+  ###################################################################### 
   observe({
     if(input$fileType=="uploadFile"){
       df <- data_set()
@@ -152,16 +191,14 @@ server <- function(input, output,session) {
       }
     
   })
-  #####################testing
-  # observeEvent(input$fileType=="sampleFile", updateRadioButtons(session, "gv1",
-  #                                                      label = "Group Variable 1",
-  #                                                      choices = c("xxxx"="x"),
-  #                                                      selected = ""))
   
   output$choose_dataset <- renderUI({
     selectInput("dataset", "Data set", as.list(data_sets))
   })
-  ####################################
+  ######################################################################
+  # Following obserfverEvent functions
+  # used to assign local values from reactive values
+  ######################################################################
   v<-reactiveValues(doViolinPlot=FALSE)
   observeEvent(input$goViolin, {
     # 0 will be coerced to FALSE
@@ -478,7 +515,10 @@ server <- function(input, output,session) {
     v55$addComment<- input$addComment
   })
   
-  
+  ######################################################################
+  # not_equalGV: compare if two group variables are the same, show warning 
+  # message if they are the same
+  ######################################################################
   not_equalGV<-function(input1, input2){
     if(strcmp(input1, input2)){
[…]
