+++ shiny-app/webinar-bio152/ui.R
@@ -1,7 +1,37 @@
+#####################################################################
+#
+# ui.R
+#     
+# 
+#
+#     This program is free software; you can redistribute it and/or
+#     modify it under the terms of the GNU General Public License,
+#     version 3, as published by the Free Software Foundation.
+#
+#     This program is distributed in the hope that it will be useful,
+#     but without any warranty; without even the implied warranty of
+#     merchantability or fitness for a particular purpose.  See the GNU
+#     General Public License, version 3, for more details.
+#
+#     A copy of the GNU General Public License, version 3, is available
+#     at http://www.r-project.org/Licenses/GPL-3
+#
+# Part of the fast-stats package
+# Contains: ui.R
+######################################################################
+
+
 library(shiny)
 library(plotly)
 library(shinydashboard)
 
+######################################################################
+# 
+# shinyUI: Create a Shiny UI handler
+# ui:
+# A user interace definition
+# 
+###################################################################### 
 shinyUI(dashboardPage(
   skin="red",
   dashboardHeader(
@@ -73,7 +103,7 @@ shinyUI(dashboardPage(
                    
                    ),
       ),
-      
+      # dataupload button
       tabItem("dataUpload",
               fluidRow(
                 box(
@@ -137,6 +167,7 @@ shinyUI(dashboardPage(
                 )
               )
       ),
+      # data visualization button
       tabItem(
         "dataVisualization",
         
