### 2022-03-14 883e3c8
update installation command
