### 2022-03-14 4bff625
update installation cmds
