+++ readme.md
@@ -65,7 +65,6 @@ install.github(“https://github.com/EdwinTh/thatssorandom")
 shiny::runApp("C:/fast-stats/shiny-app/webinar-aug20")
 shiny::runApp("C:/fast-stats/shiny-app/webinar-dec20")
 shiny::runApp("C:/fast-stats/shiny-app/webinar-bio152")
-   
 ```
 
 # Contributions
