+++ shiny-app/webinar-dec20/ui.R
@@ -5,16 +5,8 @@
 # 
 #
 #     This program is free software; you can redistribute it and/or
-#     modify it under the terms of the GNU General Public License,
-#     version 3, as published by the Free Software Foundation.
-#
-#     This program is distributed in the hope that it will be useful,
-#     but without any warranty; without even the implied warranty of
-#     merchantability or fitness for a particular purpose.  See the GNU
-#     General Public License, version 3, for more details.
-#
-#     A copy of the GNU General Public License, version 3, is available
-#     at http://www.r-project.org/Licenses/GPL-3
+#     modify it under the terms of the MIT License:
+#     https://github.com/crsl4/fast-stats/blob/master/LICENSE
 #
 # Part of the fast-stats package
 # Contains: ui.R
@@ -27,9 +19,9 @@ library(shinyjs)
 
 ######################################################################
 # 
-# shinyUI: Create a Shiny UI handler
-# ui:
-# A user interace definition
+# shinyUI: Create a Shiny UI handler (see tutorial: https://shiny.rstudio.com/tutorial/)
+# ui: user interface. This function will define the format of the 
+# screen (what the user sees when opening the web app)
 # 
 ###################################################################### 
 shinyUI(dashboardPage(
