+++ DOCS.md
@@ -1,5 +1,14 @@
 # WI Fast Stats Documentation
 
+The documentation is divided in four sections:
+
+- [Website documentation](#wi-fast-stats-website)
+- [Web apps documentation](#web-apps-documentation)
+- [Most common errors and warnings](#common-errors-and-warnings)
+- [Using WI Fast Stats to learn to code](#learning-to-code)
+
+## WI Fast Stats website
+
 WI Fast Stats is an animated website which serves as a medium to a collection of R-developed web apps that provide Data Visualization and Data Analysis tools for WI Fast Plants data. 
 
 The main website consists of 6 sections:
@@ -10,32 +19,33 @@ The main website consists of 6 sections:
 * [Source Code](#source-code)
 * [FAQ-website](#faq-website)
 
-## Home
-The Home page is a collection of animation using WI Fast Plants photos.
+### Home
+The Home page is a collection of  WI Fast Plants photos.
 ![Home-web](https://github.com/crsl4/fast-stats/blob/master/figures/home_web.png?raw=true)
 
-## About
+### About
 The About page shows the general information and features of the web apps.
 ![about-web](https://github.com/crsl4/fast-stats/blob/master/figures/about_web.png?raw=true)
 
-## Web Apps
+### Web Apps
 The Web Apps page lists all the web apps and their links.
 ![web-apps-web](https://github.com/crsl4/fast-stats/blob/master/figures/webapps_web.png?raw=true)
 
-## Webinars
-The Webinars page displays a timeline for all the WI Fast Plant webinars held so far. Each corresponding to a web app listed [above](#web-apps). More webinars will happen in the future and they will be added accordingly alongside their corresponding web app.
+### Webinars
+The developers of WI Fast Stats work closely with WI Fast Plants to organize webinars that include biological concepts and statistical concepts.
+The Webinars page displays a timeline for all the WI Fast Plant webinars held so far each corresponding to a web app listed [above](#web-apps). More webinars will happen in the future and they will be added accordingly alongside their corresponding web app.
 ![webinars](https://github.com/crsl4/fast-stats/blob/master/figures/timeline.png?raw=true)
 
-## Source Code
-The Source Code page presents the link to the Github repository.
+### Source Code
+The Source Code page presents the link to the Github repository. By providing the source code, we hope that WI Fast Stats could also serve as a entry way into programming for some audiences. You can learn more in the [Learning to Code section](#learning-to-code).
 ![source-code](https://github.com/crsl4/fast-stats/blob/master/figures/sourcecode_web.png?raw=true)
 
-## FAQ-website
+### FAQ-website
 As intended, the FAQ section contains questions and answers related to web apps and webinars. 
 
 ![faq-web](https://github.com/crsl4/fast-stats/blob/master/figures/faq_web.png?raw=true)
 
-# Web Apps Documentation
+## Web Apps Documentation
 
 Launching the app brings up the Home tab. The Home tab is basically a landing page that gives a brief introduction to the app and includes relevant information for the corresponding webinars. The following figure shows the basic Home tab.
 
@@ -47,26 +57,30 @@ The web app consists of four main tabs:
 
 ![Home](https://github.com/crsl4/fast-stats/blob/master/figures/home.png?raw=true) 
 
-## Data Upload
+### Data Upload
 
-Click Data Upload in the sidepanel to upload a data file. The dataset should be in csv format and it will be uploaded using the Data Upload box. Data should be numeric or factored and should not contain any NULL/NaN/NA values. The file should be no larger than 10MB.
[…]
