+++ notebooks/cotyledon-data-analysis.Rmd
@@ -10,13 +10,263 @@ library(ggplot2)
 library(car)
 ```
 
-# Initial data: `Polycot-data-from-NY.csv`
+
+# Data: `Polycot-data-2020.csv` focusing on chi-square test (webinar)
 
 ### Reading the data
 
 Reading in the data (had to remove manually the '#' sign from the column and save as csv file):
 ```{r}
-dat = read.table("data/Polycot-data-from-NY.csv", sep=",",header=TRUE)
+dat = read.table("../data/Polycot-data-2020.csv", sep=",",header=TRUE)
+str(dat)
+head(dat)
+summary(dat)
+dat = within(dat, Generation<-factor(Generation, levels=c("P","O")))
+```
+
+Creating categorical variable: dycots/polycots
+
+```{r}
+dat2 = data.frame(generation=dat$Generation)
+dat2$cots = "dicot"
+dat2$cots[dat$Cotyledons > 2] = "polycot"
+dat2 = within(dat2, cots<-factor(cots))
+summary(dat2)
+#write.table(dat2,file="data/polycot-data-2020-categories.csv", row.names = FALSE, sep=",")
+
+dt = table(dat2$cots, dat2$generation)
+```
+
+### Data visualization
+```{r}
+library(graphics)
+mosaicplot(dt, shade = TRUE, las=2, main="Cotelydons")
+
+library(ggmosaic)
+ggplot(data = dat2) +
+  geom_mosaic(aes(x=product(cots), fill=generation))
+```
+
+### Data analysis
+```{r}
+dt
+ct = chisq.test(dt)
+
+obs = ct$observed
+exp = ct$expected
+
+df = rbind(obs[1,],exp[1,],obs[2,],exp[2,],colSums(obs))
+rownames(df) = c("dycot observed", "dycot expected", "polycot observed", "polycot expected", "total")
+
+r = rowSums(obs)
+ss = sum(r)
+totals = c(r[1],NA,r[2],NA,ss)
+cbind(df,totals)
+ct$statistic
+ct$p.value
+
+## expected values
+2437*2556/5585
+2437*3029/5585
+3148*2556/5585
+3148*3029/5585
+```
+
+
+
+# Data: `Polycot-data-2020.csv` (topics not covered in the webinar)
+
+The code of this section contains plots and statistical tests that were not included in the final version of the webinar. For the webinar, we decided to focus on the chi-square test. Here, we fit a t test, a Levene test (for equality of variances) and a Wilcoxon-Mann-Whitney test which is similar to the t test when normality is violated. We also fit an ANOVA test and its alternative when normality is not met, the Kruskal-Wallis rank sum test.
+
+
+### Reading the data
+
+Reading in the data (had to remove manually the '#' sign from the column and save as csv file):
+```{r}
+dat = read.table("../data/Polycot-data-2020.csv", sep=",",header=TRUE)
+str(dat)
+head(dat)
+summary(dat)
+dat = within(dat, Generation<-factor(Generation, levels=c("P","O")))
+```
+
+### Data visualization
+
+```{r, echo=FALSE}
+p1 = ggplot(dat, aes(Generation, fill=Generation))+geom_bar()+
+  theme(
+        plot.title = element_text(hjust=0.5, size=rel(1.8)),
+        axis.title.x = element_text(size=rel(1.8)),
+        axis.title.y = element_text(size=rel(1.8), angle=90, vjust=0.5, hjust=0.5),
+        axis.text.x = element_text(colour="grey", size=rel(1.5), angle=0, hjust=.5, vjust=.5, face="plain"),
+        axis.text.y = element_text(colour="grey", size=rel(1.5), angle=0, hjust=.5, vjust=.5, face="plain"),
[…]
