### 2022-03-20 crsl4/fast-stats#52 @crsl4
We thank the reviewers for such helpful comments that have greatly improved the web apps and the manuscript. All of our changes are in the [review branch](https://github.com/crsl4/fast-stats/tree/review) which we will merge to master after approval by the reviewers and editorial team.

# Review 1

Functionality and Documentation:
- We double checked all the steps in the `readme` file to run the web apps locally, and provided more details on how to install `thatssorandom` and how to run the `shiny::runApp` command
- We have added more comments to the code and we have added a new section on the documentation for young audiences interested in coding
- We modified the interface so that only categorical variables appear under "Group variable" and only quantitative variables appear under "quantity"
- For the `webinar-aug20` sample data, we make `plant ID` a group variable. We found that for the other web apps, it did not make sense to change `plant ID` as categorical variable because some plots became not very intepretable, so we decided to keep them as quantitiative variables.

Software paper
- We added the DOI to me Williams and Hill (1986) paper
- We now make reference to the "Data Upload" section in the paper instead of the "Data Summary"
