### 2022-03-20 #159 @crsl4
@JasonJWilliamsNY , Yizhou and I have finished working on the reviewers' recommendations. All of our changes are in the [review branch](https://github.com/crsl4/fast-stats/tree/review) which we will merge to master after approval by the reviewers and editorial team. You can find all the responses to reviewers (same that were commented in the [JOSS review 1 issue](https://github.com/crsl4/fast-stats/issues/52) and [JOSS review 2 issue](https://github.com/crsl4/fast-stats/issues/53)) in the [response2reviews.md](https://github.com/crsl4/fast-stats/blob/review/ms/response2reviews.md) file.
