### 2022-04-20 #159 @crsl4
Thanks @JasonJWilliamsNY , I added the zenodo DOI in the readme file: https://github.com/crsl4/fast-stats
Is this what you need?
