### 2022-04-21 #159 @labarba
@crsl4 – I see that the tagged version in your repo is `[v2.1](https://github.com/crsl4/fast-stats/releases/tag/v2.1)` but here we see `v2.0`. Which one corresponds to the final revisions made for this JOSE paper?
