### 2022-04-21 #159 @labarba
@JasonJWilliamsNY – you run `@whedon set <doi> as archive` to get that into the metadata of the paper, before recommending acceptance
