### 2022-04-22 #159 @JasonJWilliamsNY
@crsl4 can you fix the invalid DOIs noted in the 'INVALID DOIs' comment from Whedon above?
