+++ ms/paper.bib
@@ -13,7 +13,7 @@ title = "https://fastplants.org/",
   pages={1385--1389},
   year={1986},
   publisher={American Association for the Advancement of Science},
-  doi = {https://doi.org/10.1126/science.232.4756.1385}
+  doi = {10.1126/science.232.4756.1385}
 }
 
 @misc{codap,
@@ -31,7 +31,7 @@ journal = {Limnology and Oceanography Letters},
 volume = {5},
 number = {6},
 pages = {379-383},
-doi = {https://doi.org/10.1002/lol2.10165},
+doi = {10.1002/lol2.10165},
 url = {https://aslopubs.onlinelibrary.wiley.com/doi/abs/10.1002/lol2.10165},
 eprint = {https://aslopubs.onlinelibrary.wiley.com/doi/pdf/10.1002/lol2.10165},
 year = {2020}
