### 2022-04-22 #159 @labarba
@crsl4 — Could you edit the metadata on the Zenodo deposit so the title and author list matches the JOSE paper?
