### 2022-04-22 #159 @crsl4
Sorry @labarba , do I do this on the Zenodo page or on github? I cannot find a place to edit this info on the zenodo site, but I could update it on the github release. Not sure if this is what you want.
