### 2022-04-22 #159 @labarba
On Zenodo, see https://help.zenodo.org

> **How can I edit the metadata of a published record?**
>
> On the record page and click the orange Edit button in the top-right corner. You will be taken to a form similar to the upload form, where you will be able to edit almost all of the record's metadata. Once you are done modifying it, click "Save" and then "Publish".
