### 2022-04-22 5f299e8
move fig; in-text citation format
for https://github.com/openjournals/jose-reviews/issues/159
