### 2022-04-22 #159 @labarba
I made a small change in PR https://github.com/crsl4/fast-stats/pull/55
- [format the citation](https://rstudio.github.io/visual-markdown-editing/citations.html) in the acknowledgement for in-text
- move the second figure, so the pagination will look better
